@@ -2,11 +2,13 @@ package com.bignerdranch.android.keepyourdaggersharp
 
 import android.os.Bundle
 import android.util.Log
+import androidx.appcompat.app.AppCompatActivity
 import com.bignerdranch.android.keepyourdaggersharp.model.NerdStore
-import dagger.android.support.DaggerAppCompatActivity
+import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
-class NerdListActivity : DaggerAppCompatActivity() {
+@AndroidEntryPoint
+class NerdListActivity : AppCompatActivity() {
 
     @Inject
     lateinit var nerdStore: NerdStore
